[#85] exit test fail
