+++ libft/Makefile
@@ -12,7 +12,6 @@
 
 NAME = libft.a
 SRCS = ft_atoi.c \
-		ft_strtol.c \
 		ft_bzero.c \
 		ft_calloc.c \
 		ft_isalnum.c \
